+++ magpylib/_src/display/traces_generic.py
@@ -280,7 +280,7 @@ def make_self_intersecting_mesh(
     return out
 
 
-def make_invalid_mesh_lines(
+def make_mesh_lines(
     obj,
     pos_orient_inds,
     mode,
@@ -294,16 +294,20 @@ def make_invalid_mesh_lines(
     style = obj.style if style is None else style
     mesh = getattr(style.mesh, mode)
     marker, line = mesh.marker, mesh.line
-    tr, vert = obj.triangles, obj.vertices
-    if mode == "disjoint":
-        subsets = obj.triangles_subsets
+    tr, vert = obj.faces, obj.vertices
+    if mode == "disconnected":
+        subsets = obj.get_faces_subsets()
         lines = get_closest_vertices(subsets, vert)
     else:
         edges = np.concatenate([tr[:, 0:2], tr[:, 1:3], tr[:, ::2]], axis=0)
         # make sure unique pairs are found regardless of order
         edges = np.sort(edges, axis=1)
         edges_uniq, edges_counts = np.unique(edges, axis=0, return_counts=True)
-        lines = vert[edges_uniq[edges_counts != 2]]
+        if mode == "open":
+            lines = vert[edges_uniq[edges_counts != 2]]
+        else:
+            lines = vert[edges_uniq]
+
     out = {}
     if lines.size != 0:
         label = f"{obj}" if label is None else label
@@ -330,10 +334,10 @@ def make_invalid_mesh_lines(
     return out
 
 
-def get_closest_vertices(triangles_subsets, vertices):
-    """Get closest pairs of points between disjoint subsets of triangles indices"""
-    nparts = len(triangles_subsets)
-    inds_subsets = [np.unique(v) for v in triangles_subsets]
+def get_closest_vertices(faces_subsets, vertices):
+    """Get closest pairs of points between disconnected subsets of faces indices"""
+    nparts = len(faces_subsets)
+    inds_subsets = [np.unique(v) for v in faces_subsets]
     closest_verts_list = []
     if nparts > 1:
         connected = [np.min(inds_subsets[0])]
@@ -376,7 +380,7 @@ def make_triangle_orientations(
     size = size if style.orientation.size is None else style.orientation.size
     offset = offset if style.orientation.offset is None else style.orientation.offset
     symbol = symbol if style.orientation.symbol is None else style.orientation.symbol
-    vertices = obj.facets if hasattr(obj, "facets") else [obj.vertices]
+    vertices = obj.mesh if hasattr(obj, "mesh") else [obj.vertices]
     traces = []
     for vert in vertices:
         vec = np.cross(vert[1] - vert[0], vert[2] - vert[1])
@@ -563,13 +567,13 @@ def make_Triangle(
     """
     vert = obj.vertices
     vec = np.cross(vert[1] - vert[0], vert[2] - vert[1])
-    triangles = np.array([[0, 1, 2]])
+    faces = np.array([[0, 1, 2]])
     # if magnetization is normal to the triangle, add a second triangle slightly above to enable
     # proper color gradient visualization. Otherwise only the middle color is shown.
     if np.all(np.cross(obj.magnetization, vec) == 0):
         epsilon = 1e-3 * vec
         vert = np.concatenate([vert - epsilon, vert + epsilon])
-        side_triangles = [
+        side_faces = [
             [0, 1, 3],
             [1, 2, 4],
             [2, 0, 5],
@@ -577,11 +581,11 @@ def make_Triangle(
             [2, 5, 4],
             [0, 3, 5],
         ]
-        triangles = np.concatenate([triangles, [[3, 4, 5]], side_triangles])
+        faces = np.concatenate([faces, [[3, 4, 5]], side_faces])
 
     style = obj.style if style is None else style
     trace = make_BaseTriangularMesh(
-        "plotly-dict", vertices=vert, triangles=triangles, color=color
+        "plotly-dict", vertices=vert, faces=faces, color=color
     )
     update_trace_name(trace, obj.__class__.__name__, "", style)
     update_magnet_mesh(
@@ -606,10 +610,10 @@ def make_TriangularMesh(
     """
     style = obj.style if style is None else style
     trace = make_BaseTriangularMesh(
-        "plotly-dict", vertices=obj.vertices, triangles=obj.triangles, color=color
+        "plotly-dict", vertices=obj.vertices, faces=obj.faces, color=color
     )
-    ntri = len(obj.triangles)
-    default_suffix = f" ({ntri} facet{'s'[:ntri^1]})"
+    ntri = len(obj.faces)
+    default_suffix = f" ({ntri} face{'s'[:ntri^1]})"
     update_trace_name(trace, obj.__class__.__name__, default_suffix, style)
     update_magnet_mesh(
         trace, mag_style=style.magnetization, magnetization=obj.magnetization
@@ -773,6 +777,8 @@ def make_mag_arrows(obj, pos_orient_inds, style, legendgroup, kwargs):
         length = obj.diameter  # Sphere
     elif isinstance(obj, magpy.misc.Triangle):
         length = np.amax(obj.vertices) - np.amin(obj.vertices)
+    elif hasattr(obj, "mesh"):
+        length = np.amax(np.ptp(obj.mesh.reshape(-1, 3), axis=0))
     elif hasattr(obj, "vertices"):
         length = np.amax(np.ptp(obj.vertices, axis=0))
     else:  # Cuboid, Cylinder, CylinderSegment
@@ -918,32 +924,32 @@ def get_generic_traces(
     orientations, positions, pos_orient_inds = get_rot_pos_from_path(
         input_obj, style.path.frames
     )
-    obj_is_disjoint = (
+    obj_is_disconnected = (
         isinstance(input_obj, TriangularMesh)
-        and style.mesh.disjoint.show
-        and not input_obj.is_connected
+        and style.mesh.disconnected.show
+        and not input_obj.check_disconnected()
     )
-    disjoint_traces = []
+    disconnected_traces = []
     for pos_orient_enum, (orient, pos) in enumerate(zip(orientations, positions)):
         if style.model3d.showdefault and make_func is not None:
-            if obj_is_disjoint:
-                tria_orig = input_obj._triangles
+            if obj_is_disconnected:
+                tria_orig = input_obj._faces
                 mag_show = style.magnetization.show
                 for tri, dis_color in zip(
-                    input_obj.triangles_subsets,
-                    cycle(style.mesh.disjoint.colorsequence),
+                    input_obj.get_faces_subsets(),
+                    cycle(style.mesh.disconnected.colorsequence),
                 ):
-                    # temporary mutate triangles from subset
-                    input_obj._triangles = tri
+                    # temporary mutate faces from subset
+                    input_obj._faces = tri
                     style.magnetization.show = False
-                    disjoint_traces.append(
+                    disconnected_traces.append(
                         make_func(
                             position=pos,
                             orientation=orient,
                             **{**make_func_kwargs, "color": dis_color},
                         )
                     )
-                input_obj._triangles = tria_orig
+                input_obj._faces = tria_orig
                 style.magnetization.show = mag_show
             else:
                 path_traces.append(
@@ -1030,10 +1036,10 @@ def get_generic_traces(
             trace["name"] = legendtext
         traces.append(trace)
 
-    if disjoint_traces:
-        nsubsets = len(input_obj.triangles_subsets)
+    if disconnected_traces:
+        nsubsets = len(input_obj.get_faces_subsets())
         for ind in range(nsubsets):
-            trace = merge_traces(*disjoint_traces[ind::nsubsets])
+            trace = merge_traces(*disconnected_traces[ind::nsubsets])
             trace.update(
                 {
                     "legendgroup": f"{legendgroup} - part_{ind+1:02d}",
@@ -1059,13 +1065,27 @@ def get_generic_traces(
             )
         )
     if isinstance(input_obj, TriangularMesh):
-        for mode in ("open", "disjoint"):
-            if mode == "open" and input_obj.is_closed:
-                continue
-            if mode == "disjoint" and input_obj.is_connected:
-                continue
+        for mode in ("grid", "open", "disconnected"):
+            if mode == "open":
+                if input_obj._status_open is None:
+                    warnings.warn(
+                        f"{input_obj!r} closed status has not been checked before attempting "
+                        "to show potential open edges, which may take a while to compute "
+                        "when the mesh has many faces, now applying operation..."
+                    )
+                if input_obj.check_open():
+                    continue
+            if mode == "disconnected":
+                if input_obj._status_disconnected is None:
+                    warnings.warn(
+                        f"{input_obj!r} connected status checked before atempting to show "
+                        "possible disconnected parts, which may take a while to compute when the "
+                        "mesh has many faces, now applying operation..."
+                    )
+                if input_obj.check_disconnected():
+                    continue
             if getattr(style.mesh, mode).show:
-                trace = make_invalid_mesh_lines(
+                trace = make_mesh_lines(
                     input_obj, pos_orient_inds, mode, label, **kwargs
                 )
                 if trace: