+++ magpylib/_src/obj_classes/class_magnet_Tetrahedron.py
@@ -102,7 +102,6 @@ class Tetrahedron(BaseMagnet):
         style=None,
         **kwargs,
     ):
-
         # instance attributes
         self.vertices = vertices
         self._object_type = "Tetrahedron"