+++ tests/test_obj_TriangularMesh.py
@@ -6,6 +6,7 @@ import pytest
 import pyvista as pv
 
 import magpylib as magpy
+from magpylib._src.exceptions import MagpylibBadUserInput
 from magpylib._src.fields.field_BH_triangularmesh import fix_trimesh_orientation
 from magpylib._src.fields.field_BH_triangularmesh import lines_end_in_trimesh
 from magpylib._src.fields.field_BH_triangularmesh import magnet_trimesh_field
@@ -33,9 +34,9 @@ def test_TriangularMesh_getBH():
     magnetization = (100, 200, 300)
     mesh3d = magpy.graphics.model3d.make_Cuboid()
     vertices = np.array([v for k, v in mesh3d["kwargs"].items() if k in "xyz"]).T
-    triangles = np.array([v for k, v in mesh3d["kwargs"].items() if k in "ijk"]).T
+    faces = np.array([v for k, v in mesh3d["kwargs"].items() if k in "ijk"]).T
 
-    triangles[0] = triangles[0][[0, 2, 1]]  # flip one triangle in wrong orientation
+    faces[0] = faces[0][[0, 2, 1]]  # flip one triangle in wrong orientation
 
     cube = magpy.magnet.Cuboid(magnetization=magnetization, dimension=dimension)
     cube.rotate_from_angax(19, (1, 2, 3))
@@ -46,32 +47,37 @@ def test_TriangularMesh_getBH():
         orientation=cube.orientation,
         magnetization=magnetization,
         vertices=vertices,
-        triangles=triangles,
-        reorient_triangles=True,
+        faces=faces,
+        reorient_faces=True,
     )
+    cube_misc_faces = cube_facet_reorient_true.to_TriangleCollection()
     cube_facet_reorient_false = magpy.magnet.TriangularMesh(
         position=cube.position,
         orientation=cube.orientation,
         magnetization=magnetization,
         vertices=vertices,
-        triangles=triangles,
-        reorient_triangles=False,
+        faces=faces,
+        reorient_faces=False,
     )
 
     vertices = np.linspace((-2, 0, 0), (2, 0, 0), 50)
     B1 = cube.getB(vertices)
     B2 = cube_facet_reorient_true.getB(vertices)
-    B3 = cube_facet_reorient_false.getB(vertices)
+    B3 = cube_misc_faces.getB(vertices)
+    B4 = cube_facet_reorient_false.getB(vertices)
 
     np.testing.assert_allclose(B1, B2)
+    np.testing.assert_allclose(B1, B3)
 
     with pytest.raises(AssertionError):
-        np.testing.assert_allclose(B1, B3)
+        np.testing.assert_allclose(B1, B4)
 
     H1 = cube.getH(vertices)
     H2 = cube_facet_reorient_true.getH(vertices)
+    H3 = cube_misc_faces.getH(vertices)
 
     np.testing.assert_allclose(H1, H2)
+    np.testing.assert_allclose(H1, H3)
 
 
 def test_TriangularMesh_getB_different_facet_shapes_mixed():
@@ -88,10 +94,10 @@ def test_TriangularMesh_getB_different_facet_shapes_mixed():
         "position": tetra.position,
         "orientation": tetra.orientation,
     }
-    tetra_facets = magpy.magnet.TriangularMesh.from_pyvista(
+    tmesh_tetra = magpy.magnet.TriangularMesh.from_pyvista(
         polydata=tetra_pv, **tetra_kwargs
     )
-    assert tetra_facets.is_reoriented
+    assert tmesh_tetra.status_reoriented is True
     cube = (
         magpy.magnet.Cuboid((111, 222, 333), (1, 1, 1))
         .move((1, 1, 1))
@@ -102,7 +108,7 @@ def test_TriangularMesh_getB_different_facet_shapes_mixed():
         "position": cube.position,
         "orientation": cube.orientation,
     }
-    cube_facets = magpy.magnet.TriangularMesh.from_pyvista(
+    tmesh_cube = magpy.magnet.TriangularMesh.from_pyvista(
         polydata=pv.Cube(), **cube_kwargs
     )
     # create a sensor of which the pixel line corsses both bodies
@@ -110,10 +116,10 @@ def test_TriangularMesh_getB_different_facet_shapes_mixed():
         [14, 65, 97], (4, 6, 9), anchor=0
     )
 
-    np.testing.assert_allclose(magpy.getB(cube, sens), magpy.getB(cube_facets, sens))
-    np.testing.assert_allclose(magpy.getB(tetra, sens), magpy.getB(tetra_facets, sens))
+    np.testing.assert_allclose(magpy.getB(cube, sens), magpy.getB(tmesh_cube, sens))
+    np.testing.assert_allclose(magpy.getB(tetra, sens), magpy.getB(tmesh_tetra, sens))
     np.testing.assert_allclose(
-        magpy.getB([tetra, cube], sens), magpy.getB([tetra_facets, cube_facets], sens)
+        magpy.getB([tetra, cube], sens), magpy.getB([tmesh_tetra, tmesh_cube], sens)
     )
 
 
@@ -122,18 +128,18 @@ def test_magnet_trimesh_func():
     mag = (111, 222, 333)
     dim = (10, 10, 10)
     cube = magpy.magnet.Cuboid(mag, dim)
-    cube_facets = magpy.magnet.TriangularMesh.from_pyvista(
+    tmesh_cube = magpy.magnet.TriangularMesh.from_pyvista(
         mag, pv.Cube(cube.position, *dim)
     )
 
     pts_inside = np.array([[0, 0, 1]])
     B0 = cube.getB(pts_inside)
-    B1 = cube_facets.getB(pts_inside)
+    B1 = tmesh_cube.getB(pts_inside)
     B2 = magnet_trimesh_field(
         "B",
         pts_inside,
         np.array([mag]),
-        np.array([cube_facets.facets]),
+        np.array([tmesh_cube.mesh]),
         in_out="inside",
     )[0]
     np.testing.assert_allclose(B0, B1)
@@ -141,26 +147,14 @@ def test_magnet_trimesh_func():
 
 
 def test_bad_triangle_indices():
-    "raise ValueError if triangles index > len(vertices)"
+    "raise ValueError if faces index > len(vertices)"
     vertices = [[0, 0, 0], [0, 0, 1], [1, 0, 0]]
-    triangles = [[1, 2, 3]]  # index 3 >= len(vertices)
-    with pytest.raises(ValueError):
+    faces = [[1, 2, 3]]  # index 3 >= len(vertices)
+    with pytest.raises(IndexError):
         magpy.magnet.TriangularMesh(
             magnetization=(0, 0, 1000),
             vertices=vertices,
-            triangles=triangles,
-        )
-
-
-def test_minimum_vertices():
-    "raise ValueError if triangles index > len(vertices)"
-    vertices = [[0, 0, 0], [0, 0, 1], [1, 0, 0]]
-    triangles = [[1, 2, 3]]
-    with pytest.raises(ValueError):
-        magpy.magnet.TriangularMesh(
-            magnetization=(0, 0, 1000),
-            vertices=vertices,
-            triangles=triangles,
+            faces=faces,
         )
 
 
@@ -175,23 +169,24 @@ def test_open_mesh():
         "z": [-1.0, -1.0, -1.0, -1.0, 1.0, 1.0, 1.0, 1.0],
     }
     vertices = np.array([v for k, v in open_mesh.items() if k in "xyz"]).T
-    triangles = np.array([v for k, v in open_mesh.items() if k in "ijk"]).T
+    faces = np.array([v for k, v in open_mesh.items() if k in "ijk"]).T
     with pytest.raises(ValueError):
         magpy.magnet.TriangularMesh(
             magnetization=(0, 0, 1000),
             vertices=vertices,
-            triangles=triangles,
-            validate_open_mesh=True,
+            faces=faces,
+            check_open="raise",
         )
 
 
-def test_disjoint_mesh():
+def test_disconnected_mesh():
     """raises Error if mesh is not connected"""
-    #  Multiple Text3D letters are disjoint
+    #  Multiple Text3D letters are disconnected
     with pytest.raises(ValueError):
         magpy.magnet.TriangularMesh.from_pyvista(
             magnetization=(0, 0, 1000),
             polydata=pv.Text3D("AB"),
+            check_disconnected="raise",
         )
 
 
@@ -240,85 +235,116 @@ def test_TriangularMesh_from_pyvista():
             get_tri_from_pv(pv.Cube())
 
 
-def test_TriangularMesh_from_facets_bad_inputs():
-    """Test from_facets classmethod bad inputs"""
+def test_TriangularMesh_from_faces_bad_inputs():
+    """Test from_faces classmethod bad inputs"""
     mag = (0, 0, 1000)
 
-    def get_tri_from_facets(facets):
-        return magpy.magnet.TriangularMesh.from_triangular_facets(
-            mag, facets, validate_open_mesh=False, reorient_triangles=False
+    def get_tri_from_triangles(trias):
+        return magpy.magnet.TriangularMesh.from_triangles(
+            mag,
+            trias,
+            check_open=False,
+            check_disconnected=False,
+            reorient_faces=False,
+        )
+
+    def get_tri_from_mesh(mesh):
+        return magpy.magnet.TriangularMesh.from_mesh(
+            mag,
+            mesh,
+            check_open=False,
+            check_disconnected=False,
+            reorient_faces=False,
         )
 
     triangle = magpy.misc.Triangle(mag, [(0, 0, 0), (1, 0, 0), (0, 1, 0)])
 
     # good element type but not array-like
     with pytest.raises(TypeError):
-        get_tri_from_facets(triangle)
-
-    # array-like but bad shape
-    with pytest.raises(ValueError):
-        get_tri_from_facets(np.array([(0, 0, 0), (1, 0, 0)]))
+        get_tri_from_triangles(triangle)
 
     # element in list has wrong type
     with pytest.raises(TypeError):
-        get_tri_from_facets(["bad_type"])
+        get_tri_from_triangles(["bad_type"])
 
-    # element in list is array like but has bad shape
+    # bad type input
+    with pytest.raises(MagpylibBadUserInput):
+        get_tri_from_mesh(1)
+
+    # bad shape input
+    msh = [((0, 0), (1, 0), (0, 1))] * 2
     with pytest.raises(ValueError):
-        get_tri_from_facets([triangle, [(0, 0, 0), (1, 0, 0)]])
+        get_tri_from_mesh(msh)
+
+    # bad shape input
+    msh = [((0, 0, 0), (1, 0, 0), (0, 1, 0), (0, 0, 1))] * 2
+    with pytest.raises(ValueError):
+        get_tri_from_mesh(msh)
 
 
-def test_TriangularMesh_from_facets_good_inputs():
-    """Test from_facets classmethod good inputs"""
+def test_TriangularMesh_from_faces_good_inputs():
+    """Test from_faces classmethod good inputs"""
     mag = (0, 0, 1000)
-
-    def get_tri_from_facets(facets, **kwargs):
-        return magpy.magnet.TriangularMesh.from_triangular_facets(mag, facets, **kwargs)
 
     # create Tetrahedron and move/orient randomly
     tetra = magpy.magnet.Tetrahedron(mag, [(0, 0, 0), (1, 0, 0), (0, 1, 0), (0, 0, 1)])
     tetra.move((3, 4, 5)).rotate_from_angax([13, 37], (1, 2, 3), anchor=0)
-    pos_orient = dict(orientation=tetra.orientation, position=tetra.position)
+    pos_ori = dict(orientation=tetra.orientation, position=tetra.position)
 
-    # copy Tetrahedron from vertices and convexhull into a TriangularMesh object
-    tetra_from_ConvexHull = magpy.magnet.TriangularMesh.from_ConvexHull_points(
-        mag, tetra.vertices, **pos_orient
-    )
+    tmesh1 = magpy.magnet.TriangularMesh.from_ConvexHull(mag, tetra.vertices, **pos_ori)
 
-    # 1-> test getB vs ConvexHull Tetrahedron
+    # from triangle list
+    trias = [magpy.misc.Triangle(mag, face) for face in tmesh1.mesh]
+    tmesh2 = magpy.magnet.TriangularMesh.from_triangles(mag, trias, **pos_ori)
+
+    # from collection
+    coll = magpy.Collection(trias)
+    tmesh3 = magpy.magnet.TriangularMesh.from_triangles(mag, coll, **pos_ori)
+
+    # from mesh
+    msh = [t.vertices for t in coll]
+    tmesh4 = magpy.magnet.TriangularMesh.from_mesh(mag, msh, **pos_ori)
+
     points = [0, 0, 0]
     B0 = tetra.getB(points)
-    B1 = tetra_from_ConvexHull.getB(points)
+    B1 = tmesh1.getB(points)
+    B2 = tmesh2.getB(points)
+    B3 = tmesh3.getB(points)
+    B4 = tmesh4.getB(points)
+
     np.testing.assert_allclose(B0, B1)
-
-    # 2-> test getB vs ConvexHull Tetrahedron faces as facets
-    facets = tetra_from_ConvexHull.facets
-    src2 = get_tri_from_facets(facets, **pos_orient)
-    B2 = src2.getB(points)
     np.testing.assert_allclose(B0, B2)
-
-    # 3-> test getB vs ConvexHull Tetrahedron faces as magpylib.misc.Triangles
-    facets = [magpy.misc.Triangle(mag, face) for face in tetra_from_ConvexHull.facets]
-    src3 = get_tri_from_facets(facets, **pos_orient)
-    B3 = src3.getB(points)
     np.testing.assert_allclose(B0, B3)
-
-    # 4-> test getB mixed input
-    facets = [magpy.misc.Triangle(mag, face) for face in tetra_from_ConvexHull.facets]
-    facets[-1] = [[0.0, 0.0, 1.0], [1.0, 0.0, 0.0], [0.0, 1.0, 0.0]]
-    src4 = get_tri_from_facets(facets, **pos_orient)
-    B4 = src4.getB(points)
     np.testing.assert_allclose(B0, B4)
+
+    # # 2-> test getB vs ConvexHull Tetrahedron faces as msh
+    # msh = tetra_from_ConvexHull.mesh
+    # src2 = get_tri_from_faces(msh, **pos_orient)
+    # B2 = src2.getB(points)
+    # np.testing.assert_allclose(B0, B2)
+
+    # # 3-> test getB vs ConvexHull Tetrahedron faces as magpylib.misc.Triangles
+    # msh = [magpy.misc.Triangle(mag, face) for face in tetra_from_ConvexHull.mesh]
+    # src3 = get_tri_from_faces(msh, **pos_orient)
+    # B3 = src3.getB(points)
+    # np.testing.assert_allclose(B0, B3)
+
+    # # 4-> test getB mixed input
+    # msh = [magpy.misc.Triangle(mag, face) for face in tetra_from_ConvexHull.mesh]
+    # msh[-1] = [[0.0, 0.0, 1.0], [1.0, 0.0, 0.0], [0.0, 1.0, 0.0]]
+    # src4 = get_tri_from_faces(msh, **pos_orient)
+    # B4 = src4.getB(points)
+    # np.testing.assert_allclose(B0, B4)
 
 
 def test_lines_ends_in_trimesh():
     "test special cases"
 
     # line point coincides with facet point
-    facets = np.array([[[0, 0, 0], [0, 1, 0], [1, 0, 0]]])
+    msh = np.array([[[0, 0, 0], [0, 1, 0], [1, 0, 0]]])
     lines = np.array([[[-1, 1, -1], [1, 0, 0]]])
 
-    assert bool(lines_end_in_trimesh(lines, facets)[0]) is True
+    assert bool(lines_end_in_trimesh(lines, msh)[0]) is True
 
 
 def test_reorient_on_closed_but_disconnected_mesh():
@@ -327,12 +353,12 @@ def test_reorient_on_closed_but_disconnected_mesh():
     s1 = pv.Sphere(theta_resolution=N, phi_resolution=N)
     s2 = pv.Sphere(theta_resolution=N, phi_resolution=N, center=(2, 0, 0))
     polydata = s1.merge(s2)
-    triangles = polydata.faces.reshape(-1, 4)[:, 1:]
+    faces = polydata.faces.reshape(-1, 4)[:, 1:]
     vertices = polydata.points
 
     # flip every 2nd normals
-    bad_triangles = triangles.copy()
-    bad_triangles[::2] = bad_triangles[::2, [0, 2, 1]]
+    bad_faces = faces.copy()
+    bad_faces[::2] = bad_faces[::2, [0, 2, 1]]
 
-    fixed_triangles = fix_trimesh_orientation(vertices, bad_triangles)
-    np.testing.assert_array_equal(triangles, fixed_triangles)
+    fixed_faces = fix_trimesh_orientation(vertices, bad_faces)
+    np.testing.assert_array_equal(faces, fixed_faces)